Batch submit
[ECOMPD2TD-1073 1707] Removal of extra flush statements and addition
of rollbacks in catch blocks instead of commits.

[ECOMPD2TD-1073 1707] Adding some needed synchronized statements to
IntegrityMonitor.

[ECOMPD2TD-1159] - Loggers should be Serializable

By making loggers Serializable, they can be included in Drools
persistent data. 'EelfLogger' and 'SystemOutLogger' can do this
trivially, but 'Logger4J' needed some additional work, because it has a
non-serializable field 'log'.

[ECOMPD2TD-000] Fix versioning of org.openecomp.policy.* dependencies

[US866186 1707] First cut of stateCheck mod and non-working JUnit

[US866186 1707] Completed the coding for the task TA1998344 which adds
a check of forward progress for dependencies.

[US866186 1707] Completed updates to IntegrityMonitor.stateCheck and
IntegrityMonitorTest which includes addition of a JUnit for stateCheck
and control of the order of JUnit execution.

[US865296] ECOMP Policy Logging Compliance, add TargetEntity and
TargetServiceName, remove unit from ElapsedTime

[US865296] ECOMP Policy Logging Compliance, set audit log statuscode to
'COMPLETE' instead of N/A

[US865296] ECOMP Policy Logging Compliance, remove time unit (seconds
and milliseconds) from ElapsedTime for logging compliance

[US866186 1707] Cleaned up IntegrityAudit JUnit tests.

[US866186 1707] IntegrityMonitor JUnit clean up

[US866186 1707] Re-added missing classes that were erroneously deleted.

[US865296] - add get/setters and inits for required log fields

Change-Id: I76ef4606ed6832ed48eaca68e72839a05c8bc3a8
Signed-off-by: Ralph Straubs <rs8887@att.com>