Batch submit
[ECOMPD2TD-000] Update version.properties for iter 41

[US866186 1707] Adding methods to IntegrityMonitor to support state
audit.

[US866186 1710] Made more progress on teh stateAudit method of
IntegritryMontor.  Completed all except the addition of properties.

[US866186 1710] Addition of
IntegrityMonitorProperties.STATE_AUDIT_INTERVAL_MS and logic to
stateAudit to determine whether or not to run

[US866186 1710] Created a JUnit for getAllForwardProgressEntity()
method and debugged and fixed it.

[US866186 1710] Making progress on the testStateAudit() JUnit.  Was able
to create forward progress entries and manually change the lastUpdated
date on an entry using a persistence update query.

[US866186 1710] First cut of stateAudit JUnit test.  Still need to test
more cases, but I got it working so far.

[US866186 1710] Completed the JUnit for stateAudit.  Added  5 new
scenarios.

Conflicts:

	version.properties

Change-Id: I7b5622a27021cf01b5d916b75824d8e51b5252eb
Signed-off-by: Ralph Straubs <rs8887@att.com>