Refactored run() in AuditThread
Sonar was complaining about a nested try/catch block in the
run method of AuditThread. It recommended moving the logic
contained within the inner try/catch to a separate method.

The nested try/catch block is necessary to capture the
exceptions properly, and have the thread continue to run.
However, separating the code as recommended by sonar
results in more readable code.

[POLICY-96]

Change-Id: I7a4eeebceeb3557841ed668c300b92d36b384292
Signed-off-by: Magnusen, Drew (dm741q) <dm741q@att.com>