Update patch set 2
Patch Set 2:

Not sure about the removal of the "isDebugEnabled" Drew.   They have an impact on performance, provided that otherwise string concatenation will be done at logging method invokation.   It can be mitigated and it will be okthough with log.debug("xx {} {}",a,b) since that won't do the concatenated until the level is checked by the logger, and not sure if FlexLogger supports that. Here perhaps is tolerable because I don't think this code will execute that often, don't think it is a good thing to change the isDebugEnabled, unless the {} mechanism is used.

Patch-set: 2