Update patch set 2
Patch Set 2:

Pam, sonar was throwing 'major' issues regarding too many nested if/else-if/else blocks in the code. All of them were caused by the "if (isDebugEnabled)" being nested in larger if/else-if blocks.

One solution I had was to create an extra method in the class that looked something like this:

private void logToDebug (Object statement) {
 if (isDebugEnabled) {
    logger.debug(statement);
 }
}

We could replace all those extra if statements with calls to that method. That would result in the removal of the excessive if statements but would still enable us to check if debugIsEnabled.

I don't know if that's an acceptable solution or not.

Patch-set: 2
Reviewer: Gerrit User 1645 <1645@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0