Update patch set 2
Patch Set 2:

I recommend to put the "isDebugEnabled" back, or alternate solution with String format that I don't think is supported by FlexLogger, easiest to put isDebugEnabled back.   String concatenation is a well known performance killer for blocks of code that executes frequently, and the concatenation that I see here is serious of many pieces of data.   We want to avoid this before the checking of the log level, so we don't want to concatenate if the level is not DEBUG which will be >99.9% of the times.  

Based on what Drew is saying, the underlying problem seems to be that sonar complains if the number of nested blocks is >N, and level N seems to be hit with the isDebug check.     Another solution, would be to create methods instead only for the debug, but go a few levels up and see if it a larger block of code can be pulled out.  This solution will make more sense if it makes the code easier to read.  Note that within eclipse and any other IDE, there are some refactor functions that could extract the method from you automatically.    

Other option is to just not deal with this issue, being a major, is not required work for this sprint to dealt with, we could then assign to Kevin as domain expert in this area, for any refactoring if needed indeed.

Patch-set: 2
Label: Code-Review=0