Tweak Integrity-Audit JUnits
This commit tweaks Integrity-Audit JUnits to allow
at least some estimate of code coverage.  It was
not possible to enable all the tests as there are
threading issues when run in the JUnit environment.
Consequently, this is about 2/3 of the actual code
coverage. Patch 2: Changes a persistence unit in DBDAOTest.

Issue-ID: POLICY-108
Change-Id: I693e56083c3ed66e020e82d5ac906ce29224bd58
Signed-off-by: Kevin McKiou <km097d@att.com>