Tweak Common Module JUnits
Patch 1: Tweaks Integrity-Audit JUnits to allow
at least some estimate of code coverage.  It was
not possible to enable all the tests as there are
threading issues when run in the JUnit environment.
Consequently, this is about 2/3 of the actual code
coverage. The estimate coverage is about 46%
Patch 2: Changes a persistence unit in DBDAOTest.
Patch3: Adds tweaks to IntegrityMonitor JUnits. The
estimated state management code coverage is 70%.

Issue-ID: POLICY-108
Change-Id: I693e56083c3ed66e020e82d5ac906ce29224bd58
Signed-off-by: Kevin McKiou <km097d@att.com>