Add AutoClose wrappers for JPA classes
Added EntityMgrCloser and EntityMgrTrans classes that will automatically
close/rollback the wrapped JPA classes.  As a result, these classes can
be used in try-with-resource statements to ensure that the JPA
EntityTransaction is rolled back and the EntityManager is closed.
Add "build" section into pom to make it consistent with other poms
in common.

Change-Id: I531f20d93a679d4bb432c1d75c1feac0e334bc44
Issue-ID: POLICY-582
Signed-off-by: Jim Hahn <jrh3@att.com>