Add AutoClose wrappers for JPA classes
Added EntityMgrCloser and EntityMgrTrans classes that will automatically
close/rollback the wrapped JPA classes.  As a result, these classes can
be used in try-with-resource statements to ensure that the JPA
EntityTransaction is rolled back and the EntityManager is closed.
Add "build" section into pom to make it consistent with other poms
in common.
Changed scope to "provided" in utils pom for javax.persistence to
prevent clashes with other projects that depend on the utils project.
Changed scope to "test" in utils pom and to "provided" in utils-test
pom for junit.
Fixed license date in EntityTransCloser.
Fixed license titles in utils & utils-test source files.
Moved files from package xxx.test.jpa to package xxx.jpa, because they
are not test-specific classes.
Include exception when logging message in ThrowablesTester, hopefully
eliminating sonar complaint.

Change-Id: I531f20d93a679d4bb432c1d75c1feac0e334bc44
Issue-ID: POLICY-582
Signed-off-by: Jim Hahn <jrh3@att.com>