Update patch set 4
Patch Set 4:

I don't have a problem with using final, most people don't use it including myself. But I see the value in it. Especially in the world of Open Source it can help ensure intention when many folks could be changing code. 
Does changing it help with the fixing of Sonar blockers? If not, let's just leave the original for now. If it does help fix the sonar blockers then let's change it.

Patch-set: 4
Label: Code-Review=0