Update patch set 1
Patch Set 1:

(3 comments)

From the code guidelines, I believe the correct implementation of a new paragraph in the javadoc is a blank line, followed by <p> at the start of new paragraph. I have replaced the </p> to align with this

"7.1.2 Paragraphs
One blank line—that is, a line containing only the aligned leading asterisk (*)—appears between paragraphs, and before the group of block tags if present. Each paragraph but the first has <p> immediately before the first word, with no space after."

Patch-set: 1
Label: Code-Review=0