Update patch set 3
Patch Set 3:

With the cleanup proposed for POLICY-1006, I understand that Dmaap* may not used, nor DME2 in favor of the UEB open source based libraries.   At least they are not in DmaaP.   If that's the case, it may be good remove them part of the POLICY-1006 work all together.

Patch-set: 3