Update patch set 4
Patch Set 4:

The reason as so that this exception could be specialized if you wanted to have a specific exception for a particular type of parameter exception. Also sometimes this exception can be thrown when say parameter values are correct but are in an illegal combination, for example if you had stream and file output and someone specified a coreect stream parameter and value but on a file outputter.

Patch-set: 4