Update patch set 4
Patch Set 4:

I'm not proposing to change the current approach to properties in existing code, the common property handling should be used for new code (initially the SDC distributor) and also in APEX, which already uses an earlier version of this.

I'll address your other comment separately.

Patch-set: 4
Label: Code-Review=0