generic jetty https server support
jetty https support in constructor, or by using
".https" when creating an http server service.

Created helper classes to manipulate the hierarchical
naming of http related properties.   The scope is
only intended to http properties, the other endpoint
ones (ueb, dmaap, noop) are left as they where.

junits for helper classes and https in the existing
HttpClientTest classes (having a mix of http and
https servers).

Change-Id: I94e8e3e4b93eb6b194657028c740b6781316c7da
Issue-ID: POLICY-940
Signed-off-by: Jorge Hernandez <jh1730@att.com>