Update patch set 5
Patch Set 5:

> Jim, have to get familiar with @Property annotation to do the work,
 > but can it do, "flattened" hierarchical property naming handling?  
 > Thing here, property name is not fixed .. ie. http.server.services.X.https
 > and http.server.services.Z.https ..
 > 
 > Point being the property names are unbound and not known a priori. 
 > So, question is if @Property supports such functionality?

If I understand the question, there's a helper class, SpecProperties that helps with that.  In the above example, you'd create: sp = SpecProperties("http.server.services."+serviceName, props).  Then you can do sp.getProperty("https.xxx"), etc.  (Or include the ".https" in the constructor, if that's preferred.)  You'd pass "sp" instead of "props" to the PropertyConfiguration subclass.  The same kind of thing is done in the pooling example, where the property names have the controller name in the prefix

Patch-set: 5