Update patch set 5
Patch Set 5: Code-Review+1

Maybe we should just go with this as it is, and I can retrofit the properties after it's merged.  I still prefer the constants to be moved out of the interfaces, particularly since this is new code, but I'm not dead set on that either.

Patch-set: 5
Label: Code-Review=+1