Update patch set 6
Patch Set 6:

I trimmed the review to just focus on strict https support, which is the main intent of the work, and removed the helper classes all together as it was a secondary theme, so left the original code in that area untouched.   The discussion about the 2 approaches proposed: common-parameters and @Property, and application on a generic way across to deal with properties merits its own discussion thread outside this review I believe.

Patch-set: 6