Fix bug in common-paramter validation service
In GroupValidationResult & GroupMapValidationResult, the setResult
method was not setting the actual status that it has recieved from the
nested group rather it was setting its own status again. Due to which
the validation status of overall ParameterGroup was coming as CLEAN even
if a nested child has reported it as INVALID.
Changed the expected validation result texts for TestValidation accordingly.

Change-Id: I5943decce39574089ca74f9bbccbfbe66d8d542c
Issue-ID: POLICY-1035
Signed-off-by: ramverma <ram.krishna.verma@ericsson.com>