Update patch set 4
Patch Set 4:

> (1 comment)

Actually, the server list is expected to have only 1 entry. because i looked at the usage, it seems, we only consider first element and not the rest. getServers().get(0). Hence based on this, ==1 seems to be correct, if the first element is empty, then it is invalid, 
Valid:
1) {"127.0.0.1"}


Invalid:
1) {"", "127.0.0.1"}
2) {" ", "127.0.0.1"}

Are there any other scenarios ?

Patch-set: 4
Label: Code-Review=0