Update patch set 2
Patch Set 2:

are we sure there is no sideeffects for this change in casablanca given the stage that we are at?   I remember at some point, the "ecomp" naming on that section, it was necessary as was needed by LF infrastructure.   I may be confused, seems a bit blurry to me, but want to make sure.  Are similar lines in the some of the parent poms?

Patch-set: 2
Reviewer: Gerrit User 477 <477@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0