Update patch set 8
Patch Set 8: -Code-Review

> (1 comment)

Jorge, your latter assessment is correct, and the former may be, as well.  Not sure there's an easy answer.  If this is not excluded, then there is no way to force it to use gson, as media-json-jackson will take precedence (randomly, as it turns out).  Got any ideas?

Patch-set: 8