Update patch set 8
Patch Set 8:

Yeah, it's tricky.  I don't have an easy answer neither.  

I would assume that the dmaap team added the jersey-media-json-jackson as a direct dependency, the jersey http client they use, more than likely would use it under the hood for json payloads.

If we remove it, if our properties configuration triggers the use of this library, I think it may be when DmaaP uses AAF.   I am concerned that this client functionality could fail in some obscure ways.   For that reason, I think the exclusion should be undone.

On the other hand, this is a nice piece of work to use gson instead, but I am afraid the solution can not be full blown used unless our dependencies does not transitively include this library.

Another issue that I see down the line is replacing the @JsonIgnore with "transient" to limit serialization in other repos.   If the json-media-json-jackson appears by virtue of transitive dependencies somewhere in the executable package.  When going and doing the replacement, I think the transient may be ignored since the default library would be the jackson and would look for the JsonIgnore.

I guess, cannot think of a good solution.

One idea is to submit this review as it provides the gson support, undo the exclusion, and postpone the rest of the work in the other repos until ONAP as a whole decides to eliminate Jackson.   I think these changes are transparent pretty much to other repos with the benefit that the gson support is there now.

I think though you may have a better handle on the issues that I described, may be I am wrong in some of my conclusions.    What are your thoughts?

Patch-set: 8
Label: Verified=0