Update patch set 2
Patch Set 2:

Up to now, I have marked fields that I don't want to serialize or deserialize as transient. I recognize that this goes beyond GSON.

It seems that three is an @Expose annotation in GSON that works the other way around, you mark what you want to serialize/deserialize.

I see that the approach you used is also suggested on this stack overflow post:
https://stackoverflow.com/questions/4802887/gson-how-to-exclude-specific-fields-from-serialization-without-annotations

I just want to ask the question, can we not just use @Expose when we need to pick the fields to serialize?

Patch-set: 2