Update patch set 2
Patch Set 2:

> Up to now, I have marked fields that I don't want to serialize or
 > deserialize as transient. I recognize that this goes beyond GSON.
 > 
 > It seems that three is an @Expose annotation in GSON that works the
 > other way around, you mark what you want to serialize/deserialize.
 > 
 > I see that the approach you used is also suggested on this stack
 > overflow post:
 > https://stackoverflow.com/questions/4802887/gson-how-to-exclude-specific-fields-from-serialization-without-annotations
 > 
 > I just want to ask the question, can we not just use @Expose when
 > we need to pick the fields to serialize?

Using "transient" seemed insufficient, as there may be cases where you want the java built-in serialization to include a field (e.g., persistence), but don't want a field serialized when exported.

As for @Expose, I believe it's an all or nothing proposition; if @Expose processing is enabled, then the annotation must be used in all classes serialized via that gson object.  With @JsonIgnore, it's easy to scan classes looking for that annotation and add a @GsonIgnore where appropriate.  However, there may be a number of classes that currently have no jackson annotations and there's no simple mechanism to find those to add @Expose annotations through-out.  Consequently, using @GsonIgnore seemed to be the approach that was least likely to cause things to be missed when converting from jackson to gson.

Patch-set: 2