Update patch set 2
Patch Set 2:

Nothing special to be honest. I'm learning on Policy FW and I opened this piece of code and I saw that it's a total mess there in terms of architecture and code (copy-paste programming for example) so now I'm trying to get more deeply in it and rework Topic Factories as mention in parent of this task.

This series is just a small fix that I did in the beginning and now I'm working further on simplifying topic factories

Patch-set: 2
Label: Code-Review=0