Update patch set 2
Patch Set 2:

Krzysztof has my agreement on reworking the Topic Factories for UEB and DMaaP, the work was proposed under POLICY-1387 and POLICY-1486 that this review points at is a subtask.   See comments in POLICY-1387.  I agree that there's a lot of copy/pasted code there, especially the DMaaP* classes that pretty much were duplicated from the UEB ones.   It should help at least with the sonar metrics with regards to code duplication.

Patch-set: 2