Fix potential race condition
init() is currently called outside of synhronized block.
This is may lead to race condition as init() creates new bus consumer
(this.consumer) which is the copied and nulled in stop() synchronized
block. This may lead to race condition if:

Thread #1   	     	     	       Thread #2

this.alive = true;
					this.alive = false
					consumerCopy = this.consumer;

this.consumer = new Consumer();

					// false
					if(consumerCopy != null) {
					    consumerCopy.close();
					} else {
					    log.warn();
					}

As a result, new BusConsumers are being created but they are never
close().

Issue-ID: POLICY-1387
Change-Id: I2eadb12ef1c4b07b9e47dc6ebc096acc713299a0
Signed-off-by: Krzysztof Opasiak <k.opasiak@samsung.com>