Update patch set 4
Patch Set 4:

> Jim, if I understand correctly, the jackson annotations JsonIgnore,
 > JsonProperty, etc .. are honored at serialization/deserialization
 > time.  Is my understanding correct?
 > 
 > I think you had in the past counterpart annotations for GsonIgnore,
 > etc .. but I think you abandoned that work.   I was wondering if
 > you could resurrect it, as for a new class coded up directly with
 > Gson, it may make more sense to just use them?   What are your
 > thoughts?
 > 
 > There's also a question from Pam to answer about right jackson
 > version, in case you missed it.

Jorge: yes, your understanding is correct, or it will be, once all of this code is put in place.

As for resurrecting the GsonXxx flavors, we should probably discuss that in the JIRA comments.  A given message stream should use one or the other, but not both.  Would you expect the stream using the GsonXxx flavors to follow the gson default behavior, or the jackson default behavior?  If the latter, then I would assume we'd want to migrate the jackson annotations to gson annotations for all classes in that stream.

I couldn't find the 2.9.6 comment, so will have to look around a bit more.

Patch-set: 4