Update patch set 4
Patch Set 4:

Thanks for the explanation, agreed we can move comments to JIRA ticket, may be we can add the conclusion to the resolution of jira ticket.  You are expert on this area so I'll follow your guidance, but it seems to me more user friendly to have extensions, new Gson annotations, named for example "@GsonJacksonStyleIgnore", "@GsonJacksonStyleProperty" that follows the Jackson behavior.  The annotation name would make it explicitly clear.   I think this will allow replacing all existing Jackson annotations and the direct dependencies to Jackson libraries.  Otherwise unless it is really well documented, I think folks could get confused as gson may be used, but still jackson annotations would be honored.   I realize though that it introduces extra work, so that if it is significant it is a drawback.   Makes sense?   I am interested in hearing your thoughts on it.

Patch-set: 4