Update patch set 6
Patch Set 6: -Code-Review

> (4 comments)

Personally, I'd create a subclass of Thread containing everything it needs, including a "stop" flag, implemented via a CountDownLatch.  Create and store the thread object when it's running.  Stop it and store null when done.  Don't bother with interrupts.  Could also override isAlive() to check that it's non-null and the thread, itself, is alive.

Patch-set: 6