Update patch set 1
Patch Set 1:

What would really be nice would be the ability to make the UEB/DMAAP decision using the effective topic. One way to do this in an upward-compatible manner would be to allow properties like 'source.topics' in addition to 'dmaap.source.topics', for example:

source.topics=DCAE-IN
...
source.topics.DCAE-IN.effectiveTopic=dmaap.com.att.dcae.DCAE-CL-EVENT

For topics defined generically like this, the 'effectiveTopic' must be defined, and the DMAAP/UEB type would need to be included, or somehow implied.

Patch-set: 1