Update patch set 1
Patch Set 1:

I think Ralph this is brought up as a side conversation not directly related to the review.  

This is my feedback: 

The ueb/dmaap consideration can be considered "almost" deprecated in practical terms.  

In ONAP for example, we use (which will fetch across the global set of endpoints for it):

PolicyEngine.manager.deliver(topic-name, Object-to-send).   

There's no distinction between UEB/DMAAP flag.   This is from the standpoint that it does not make any sense to talk to the same topic in a given lab deployment using ueb and dmaap, unless it is an obscure, bizarre lab configuration that somebody perhaps may want to support. 
   
The way I see this evolving when somebody has the cycles to do code cleanup is that one of ueb/dmaap should be integrated into one.  Actually the "dmaap" flag will try to use cambria libraries whenever possible.

I think effectively the 2 argument deliver(topic-name, object-to-send), and being able to override the "real" topic name with the optional effective topic, gets rid of the need to provision and expand topic names in the template, it is now a configuration problem instead, rather than being both a drl template and configuration problem.

Patch-set: 1