Update patch set 1
Patch Set 1:

Jim, just lightly touching in your first comment, embedding ${X} information in the drl template as variables is pretty much what brmsgw does, and that approach I see it problematic at many levels, since drl in theory should be a design time artifact.  We can have conversation offline if you want to on this, but I think you're OK with the changes based on your last comment in the review.   As a note, taking out junits and reformatting the review is not sizable, around 100 lines.

Patch-set: 1