Update patch set 1
Patch Set 1:

> (1 comment)

still need additional test cases per the comment on line 83.

Regarding line 84, since you really want to verify that the reference is the same, that's fine.  However, in general, should use assertEquals() any time it doesn't matter if the reference is the same, as it provides more functionality in Eclipse if there's a mismatch.  (Eclipse provides a pop-up highlighting the differences.  That's why I sometimes use assertEquals(xxx.toString(), yyy.toString()) instead of assertEquals(xxx,yyy), because it highlights the mismatching fields.)

Patch-set: 1