Add common coder classes
Added coder classes as a common utility that can be used to encode
and decode json data.  Currently, it uses gson, but that can be
changed.

Separated the Coder interface from implementation.

Used assertSame to compare references.
Used numbers instead of strings so don't have to deal with escaping quotes
in the test cases.

Change-Id: Iea9001d2adc5a3150b3367619b5f43e96b5f8b41
Issue-ID: POLICY-1444
Signed-off-by: Jim Hahn <jrh3@att.com>