Update patch set 4
Patch Set 4:

> (3 comments)
 > 
 > Jorge, I am not sure that these messages will come under drools
 > session or not. Anyways we can achieve serialization using
 > Coder/Decoder work that Jim is doing (using gson).
 > @Jim, is that correct?

Session persistence is all done via the Serialization interface and we have no control over it, as it's hidden within drools.  The question is: will this ever have to be persisted by the drools engine?  I kind of doubt it, but not sure.  If it is persisted, then yes, it must implement the Serialization interface.

Patch-set: 4