Update patch set 1
Patch Set 1:

> Is this logger compliant with ONAP logging standards?

I'm not too familiar with the logging standards after Amsterdam so I will have to get back to you on that. However the logger that this code uses is the standard network log that has been in use since Amsterdam so if the standard network logger is compliant than it should be fine.

Patch-set: 1