Update patch set 1
Patch Set 1:

> A few logistics things that need to be taken care of:
 > 
 > a) add junits,
 > 
 > b) a plan to have coordinated changes submitted too as drools-pdp
 > and drools-applications. If you can provide feedback.
 > 
 > As a side note, I think this class should be deleted or
 > consolidated with NetLoggerUtil.java:
 > 
 > https://git.onap.org/policy/common/tree/utils/src/main/java/org/onap/policy/common/utils/slf4j/LoggerFactoryWrapper.java
 > as it does not have relevance.
 > 
 > Question:
 > 
 > Does this cover all sources, specifically Noop sources?

a) Sure, will add in the next review
b) Part 1 is to have this submitted as it won't disrupt any current features in the platform. Part 2 would be to submit two pull requests simultaneously in drools-pdp and drools-applications to migrate all features to use common's OrderedService. Once those are merged part 3 is to add the mdc filter feature. I understand these changes are quite disruptive and am open to suggestions for another approach. I found the trade offs the most fitting for what the feature needs, which is a way to intercept messages before they are logged so they can be placed in the MDC to be used by logback. Ideally the least disruptive change would have this be done at the controller level but that would introduce duplicate messages if 2 controllers ingest a message from a shared topic.

On the LoggerFactoryWrapper, I agree and will proceed with the changes.

On your question, when I was looking through the new topic code it appeared that both NoopTopicSource and NoopTopicSink call the NoopTopicEndpoint's io() method, which would mean that noop sources would be covered. Just an fyi, one thing that seemed like a minor bug to me is that by calling io() for both the noop source/sink, the method always logs the message as an "OUT" message rather than an "IN" message if it is a source topic. But perhaps I am missing some context.

Patch-set: 1