Add code to allocate server ports
Added code to find available server ports.
Also added a trust manager that always trusts certificates.

Made trust manager private so we can allocate new ones in the future,
if we decide it isn't safe to re-use the same one over and over.

Modified test code to allocate a port rather than assuming 8180 was
not in use.  Also modified retry parameters in other tests.

Added another allocPort() that takes an IP address and then modified
the other two to use that so the code does not have to be duplicated,
while providing more functionality.

Change-Id: I08cf18d923af46b1310d8496498bb1d4a0690b6e
Issue-ID: POLICY-1542
Signed-off-by: Jim Hahn <jrh3@att.com>