Update patch set 1
Patch Set 1:

> about encrypt password with salt, is salt similar to secretKey in
 > our code? Here is an article about encrypt password with salt:
 > https://examples.javacodegeeks.com/core-java/security/encrypt-decrypt-with-salt/

No, salt is not the same as the secret key.  The salt causes a new encrypted string to be generated every time "encrypt" is called, thus making it more difficult for someone to guess the key.  This article is better:
https://www.digizol.com/2009/10/java-encrypt-decrypt-jce-salt.html

However, a simple way to do it is to generate a fixed-length salt and preppend or append it to the text to be encrypted.  After decrypting, just strip off the salt.

Patch-set: 1