Update patch set 3
Patch Set 3:

One idea is that you change the overloaded method with the allocPort(String hostname ..) interface to pass InetAddress instead (it may be better to give that responsability to client, as to get the String hostname properly the caller may have to go through some java calls to do it properly).   Changing the interface to that, and you pass null in that interface underlying java code will resolve without a dns look up.   Let me know what you think.

Patch-set: 3