Update patch set 2
Patch Set 2:

(1 comment)

Yeah, I can see it useful in junits.  

That would be nice if you can add a new method to support it.   Not sure if there are some "suffix terminology" such as the *IfAbsent as in the concurrent hashtable that could be reused to distinguish between the 2 flavors.   Whatever you decide is good with me.   Thank you!

Patch-set: 2