Update patch set 5
Patch Set 5:

> Hi Jim,
 > 
 > Just wondering where this is going to be placed? The DaoImpl code
 > hides the entity and transaction handling for the JPA.

Liam: That's just it, the DaoImpl should neither use nor hide transaction handling.  The invoker should be controlling the transaction, as it will want to do queries, updates, and creates, all within a single transaction.

In addition, the DaoImpl should not create the Entity Manager object, either; that should be done by the factory class, which will then pass that object into the constructor for DaoImpl.  It should be possible to have separate Impl classes for each part of the DB to be manipulated, and then aggregate these classes inside a single Impl, sharing the same Entity Manager across all of the objects.  This also implies that the "init(params)" method should be removed from the PfDao interface.

Patch-set: 5